优化图片管理删除体验

+++ A/c/SysController.php
@@ -193,7 +193,9 @@ class SysController extends CommonController
 			
 			if(M('pictures')->delete('id in('.$data.')')){
 				foreach($pictures as $v){
-					unlink('.'.$v['litpic']);
+					if(file_exists('.'.$v['litpic'])){
+						unlink('.'.$v['litpic']);
+					}
 				}
 				
 				JsonReturn(array('code'=>0,'msg'=>'批量删除成功！'));
